// KivuIMSFrontendContainerApp/app/components/Profile/enums/profileEnums.ts
export enum StoreStatus {
  PENDING = "PENDING",
  ACTIVE = "ACTIVE",
  SUSPENDED = "SUSPENDED",
  DELETED = "DELETED"
}

export enum UserRole {
  STORE_ADMIN = "STORE_ADMIN",
  INVENTORY_CLERK = "INVENTORY_CLERK",
  WAREHOUSE_MANAGER = "WAREHOUSE_MANAGER",
  SALES_PROCUREMENT = "SALES_PROCUREMENT",
  BUYER = "BUYER",
  ANALYST = "ANALYST"
}

export enum AccountStatus {
  INVITED = "INVITED",
  ACTIVE = "ACTIVE",
  SUSPENDED = "SUSPENDED",
  DELETED = "DELETED"
}